+++ src/App.js
@@ -10,6 +10,7 @@ import Category from './Containers/Category/Category';
 import ProductDetails from './Containers/ProductDetails/ProductDetails';
 import SignIn from './Containers/SignIn/SignIn'
 import SignUp from './Containers/SignUp/SignUp';
+import Cart from './Containers/Cart/Cart';
 
 const App = () => {
   let routes = (
@@ -17,6 +18,7 @@ const App = () => {
       <Route path="/home" component={Home} />
       <Route path="/signin" component={SignIn} />
       <Route path="/signup" component={SignUp} />
+      <Route path="/cart" component={Cart} />
       <Route path="/product/:url/:id" component={ProductDetails} />
       <Route path="/:category/:subCategory" component={Category} />
       <Route path="/" component={Home} />
