+++ src/Components/Auth/SignIn/Login.css
@@ -16,7 +16,7 @@
 }
 
 .signup-form-wrapper {
-    width: 50%;
+    width: 100%;
     display: flex;
     justify-content: center;
     margin: 20px auto;
@@ -43,4 +43,10 @@
 
 .signup button {
     width: 10%;
+}
+
+@media (min-width: 480px) {
+    .signup-form-wrapper {
+        width: 50%;
+    }
 }
