+++ src/Components/PopUp/PopUp.css
@@ -1,0 +1,73 @@
+.popup-content {
+    position: absolute;
+    background-color: black;
+    color: white;
+    display: block;
+    align-items: center;
+    text-align: center;
+    padding: 10px 50px;
+    z-index: 200;
+    margin-left: -200px;
+    transition: all 0.3s ease;
+}
+
+.open-popup {
+    transform: translateY(90%);
+}
+
+.close-popup {
+    transform: translateY(-100%);
+}
+
+@media (min-width: 480px) {
+    .popup-content {
+        position: fixed;
+        background-color: black;
+        color: white;
+        display: block;
+        align-items: center;
+        text-align: center;
+        padding: 10px 50px;
+        z-index: 200;
+        max-height: 70%;
+        margin-left: -326px;
+        overflow-y: auto;
+        transition: all 0.3s ease;
+    }
+    
+    .open-popup {
+        top: 0;
+        transform: translateY(100px);
+    }
+    
+    .close-popup {
+        top:0;
+        transform: translateY(-100%);
+    }
+
+    .popup-content::-webkit-scrollbar-track {
+        border: 1px solid #000;
+        padding: 2px 0;
+        background-color: rgba(255, 255, 255, 0);
+    }
+      
+    .popup-content::-webkit-scrollbar {
+        width: 10px;
+    }
+      
+    .popup-content::-webkit-scrollbar-thumb {
+        border-radius: 10px;
+        box-shadow: inset 0 0 6px rgba(0,0,0,.3);
+        background-color: #737272;
+        border: 1px solid #000;
+    }
+
+    .popup-content::-webkit-scrollbar-thumb:hover {
+        box-shadow: inset 0 0 6px rgba(66, 66, 66, 0.3);
+    }
+
+    .popup-content {
+        scrollbar-width: thin;          /* "auto" or "thin" */
+        scrollbar-color: rgb(82, 82, 82) rgba(255, 255, 255, 0);   /* scroll thumb and track */
+    }
+}
