+++ src/Containers/Cart/Cart.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import './Cart.css';
+
+const Cart = () => {
+
+    return (
+        <>
+            carrito
+        </>
+    );
+};
+
+export default Cart;
