+++ src/index.css
@@ -26,3 +26,29 @@ button, input[type="submit"], input[type="reset"] {
 input:focus {
   outline: none;
 }
+
+body::-webkit-scrollbar-track {
+  border: 1px solid #000;
+  padding: 2px 0;
+  background-color: rgba(255, 255, 255, 0);
+}
+
+body::-webkit-scrollbar {
+  width: 10px;
+}
+
+body::-webkit-scrollbar-thumb {
+  border-radius: 10px;
+  box-shadow: inset 0 0 6px rgba(0,0,0,.3);
+  background-color: #737272;
+  border: 1px solid #000;
+}
+
+body::-webkit-scrollbar-thumb:hover {
+  box-shadow: inset 0 0 6px rgba(66, 66, 66, 0.3);
+}
+
+body {
+  scrollbar-width: thin;          /* "auto" or "thin" */
+  scrollbar-color: rgb(82, 82, 82) rgba(255, 255, 255, 0);   /* scroll thumb and track */
+}
